requirements.txt
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
certifi==2024.6.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==2.6.0
crispy-bootstrap4==2024.1
crispy-tailwind==1.0.3
Django==5.0.6
django-browser-reload==1.13.0
django-crispy-forms==2.2
django-modeltranslation==0.19.5
django-tailwind==3.8.0
idna==3.7
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==2.0.0
packaging==24.1
pandas==2.2.2
plotly==5.22.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
six==1.16.0
sqlparse==0.5.0
tenacity==8.5.0
text-unidecode==1.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
